# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "Bootcamp 2025 API",
    "version": "1.0.0",
    "description": "API para gerenciamento de atividades e usuários do Bootcamp"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "security": [],
        "summary": "Registrar novo usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "João Silva" },
                  "email": { "type": "string", "example": "joao@email.com" },
                  "password": { "type": "string", "example": "Senha123@" },
                  "cpf": { "type": "string", "example": "123.456.789-00" }
                },
                "required": ["name", "email", "password", "cpf"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": {
                  "message": "Usuário criado com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Informe os campos obrigatórios corretamente"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": "O e-mail ou CPF informado já pertence a outro usuário"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sign-in": {
      "post": {
        "security": [],
        "summary": "Login do usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "joao@email.com" },
                  "password": { "type": "string", "example": "Senha123@" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": "3f885f64-5717-4562-b3fc-2c963f68afab",
                    "name": "João Silva",
                    "email": "joao@email.com",
                    "cpf": "string",
                    "avatar": "string",
                    "xp": 0,
                    "level": 0,
                    "achievements": [
                      {
                        "name": "string",
                        "criterion": "string"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Informe os campos obrigatórios corretamente"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Senha incorreta"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Usuário não encontrado"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Obter dados do usuário",
        "tags": ["Usuários"],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "example": {
                  "id": "3f885f64-5717-4562-b3fc-2c963f68afab",
                  "name": "João Silva",
                  "email": "joao@email.com",
                  "xp": 100,
                  "level": 2
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/user/preferences": {
      "get": {
        "summary": "Obter preferências do usuário",
        "tags": ["Usuários"],
        "responses": {
          "200": {
            "description": "Lista de preferências",
            "content": {
              "application/json": {
                "example": [
                  {
                    "typeId": "c7f42b7b-4b32-453c-902d-33baae755607",
                    "typeName": "Yoga"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Preferências não encontradas"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/user/preferences/define": {
      "post": {
        "summary": "Definir interesses do usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "typeId": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "message": "Preferências atualizadas com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Um ou mais ID informados são inválidos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/user/avatar": {
      "put": {
        "summary": "Atualizar avatar do usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar atualizado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Avatar atualizado com sucesso",
                  "avatarUrl": "http://example.com/avatar.jpg"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Arquivo inválido ou não enviado"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "example": {
                  "error": "Arquivo muito grande (tamanho máximo: 5MB)"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao processar upload do avatar"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "summary": "Atualizar dados do usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dados atualizados",
            "content": {
              "application/json": {
                "example": {
                  "message": "Dados atualizados com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Dados inválidos fornecidos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": "O e-mail informado já está em uso por outro usuário"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao atualizar dados do usuário"
                }
              }
            }
          }
        }
      }
    },
    "/user/deactivate": {
      "delete": {
        "summary": "Desativar conta",
        "tags": ["Usuários"],
        "responses": {
          "200": {
            "description": "Conta desativada",
            "content": {
              "application/json": {
                "example": {
                  "message": "Conta desativada com sucesso"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Operação não permitida"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro ao desativar conta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}